Move bundled skill loading into shared module
